Edits to greatly improve performance
Added classes to store bones that are supposed to jiggle instead of making it run through all bones and objects on each frame.
Unfortunately, there's no handler that can run the function before animation plays so there's a new button 'Initialize bones' that you need to run if you've changed the number of jiggle bones. 
This change greatly improves performance on complex scenes with a lot of bones/objects.